chore: split the parsers in separate modules
